Reorganize all ionosphere classes into one file and rename module files to be lowercase.
